[bugfix] Remove ambiguous overload of abs in divide_by_i
The divide_by_i function fails to compile with an ambiguous
overload of the function abs(int). The generic way of fixing
this seems to be to add a `using std::abs` statement that
enables ADL to find the correct overload. In this specific use
case however, the whole abs-call seems to be superfluous as
it checks `abs(k-1)` within a `k>1`-if-clause.
I have therefore removed the abs-call altogether.